redux store setup completed
