+++ src/redux/baseApi.ts
@@ -1,0 +1,9 @@
+import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react"
+
+
+
+export const baseApi = createApi({
+    reducerPath: "baseApi",
+    baseQuery: fetchBaseQuery( { baseUrl: "http://localhost/3000/ap1/v1"}),
+    endpoints: () => ({})
+})
