+++ src/redux/store.ts
@@ -1,0 +1,18 @@
+import { configureStore } from '@reduxjs/toolkit'
+import { baseApi } from './baseApi'
+import { setupListeners } from '@reduxjs/toolkit/query'
+
+export const store = configureStore({
+  reducer: {
+    [baseApi.reducerPath]: baseApi.reducer
+  },
+  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(baseApi.middleware)
+})
+
+
+setupListeners(store.dispatch)
+
+// Infer the `RootState` and `AppDispatch` types from the store itself
+export type RootState = ReturnType<typeof store.getState>
+// Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
+export type AppDispatch = typeof store.dispatch
